Moved `ReadModelFilter` type util to `ReadModelRepository` file
